docker/web/pyproject.toml
[tool.poetry]
name = "django_start"
version = "0.1.0"
description = "A starting Django project environment"
authors = ["Stephen Kelley <stevek@virtustate.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
numpy = "^1.19.0"
pandas = "^1.0.5" 
psycopg2 = "^2.8.5"
scikit-learn = "^0.23.1"
django_redis = "^4.12.1"
snowflake_sqlalchemy = "^1.2.3"
matplotlib = "^3.3.0"
plotly = "^4.9.0"
django-bootstrap-breadcrumbs = "^0.9.2"
django_bootstrap4 = "^2.2.0"
django_sass_processor = "^0.8"
django_analytical = "^2.6.0"
django_fontawesome_5 = "^1.0.18"
django_libsass = "^0.8"
django_environ = "^0.4.5"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
flake8 = "^3.8.3"
black = "^19.10b0"
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
